PF-1465 - Add support for delegate plugin configuration
The MultiFile plugin can now be configured with a list of plugins, their priorities, and their settings.

It would be great if MultiFile could automagically figure this out, but that is finicky, since the plugin might be running in different contexts (on an app server, launched from the tester, or from withing the FVHFS)

For now, an explicit configuration JSON setting will need to suffice.